FATHOM5 updates to package haversine
- Updated travis.yml to point to the FATHOM5 build
- Updated README to reflect the API changes introduced in the commit
- Updated doc.go to reflect the API changes
- Updated haversine.go to make the API more atomic.  The intent here
was to return only a single value in each call to the Distance function.
Then to add a few more functions to get the distance in the units
desired by the client.
- Moved the tests into the main repo folder.  More importantly added
math.Trunc() on returned values so that accurate Haversine conversions
were not failing due to floating point accumulation errors.